Stop supporting Pydantic-V1 JSON-Schema generation (ref #802)
